requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
Django==3.0
django-filter==2.2.0
django-nose==1.4.6
djangorestframework==3.10.3
entrypoints==0.3
factory-boy==2.12.0
Faker==3.0.0
flake8==3.7.9
flake8-django==0.0.4
idna==2.8
importlib-metadata==1.2.0
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
nose==1.3.7
packaging==19.2
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.1
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
